Swap tutorial3 and tutorial3-completed
