+++ android/tutorials/tutorial3-completed/src/main/java/com/uber/rib/root/logged_out/LoggedOutInteractor.java
@@ -37,13 +37,13 @@ public class LoggedOutInteractor
   protected void didBecomeActive(@Nullable Bundle savedInstanceState) {
     super.didBecomeActive(savedInstanceState);
     presenter
-        .loginName()
+        .playerNames()
         .subscribe(
             new Consumer<Pair<String, String>>() {
               @Override
               public void accept(Pair<String, String> names) throws Exception {
                 if (!isEmpty(names.first) && !isEmpty(names.second)) {
-                  listener.login(names.first, names.second);
+                  listener.requestLogin(names.first, names.second);
                 }
               }
             });
@@ -56,10 +56,11 @@ public class LoggedOutInteractor
   /** Presenter interface implemented by this RIB's view. */
   interface LoggedOutPresenter {
 
-    Observable<Pair<String, String>> loginName();
+    Observable<Pair<String, String>> playerNames();
   }
 
   public interface Listener {
-    void login(String userNameA, String userNameB);
+
+    void requestLogin(String playerOne, String playerTwo);
   }
 }
