+++ android/tutorials/tutorial3-completed/src/test/java/com/uber/rib/root/logged_in/off_game/OffGameInteractorTest.java
@@ -16,15 +16,20 @@
 package com.uber.rib.root.logged_in.off_game;
 
 import com.uber.rib.core.RibTestBasePlaceholder;
+import com.uber.rib.root.logged_in.ScoreStream;
 import org.junit.Before;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 
 public class OffGameInteractorTest extends RibTestBasePlaceholder {
 
+  private final String playerOne = "playerOne";
+  private final String playerTwo = "playerTwo";
+
   @Mock OffGameInteractor.Listener listener;
   @Mock OffGameInteractor.OffGamePresenter presenter;
   @Mock OffGameRouter router;
+  @Mock ScoreStream scoreStream;
 
   private OffGameInteractor interactor;
 
@@ -32,6 +37,7 @@ public class OffGameInteractorTest extends RibTestBasePlaceholder {
   public void setup() {
     MockitoAnnotations.initMocks(this);
 
-    interactor = TestOffGameInteractor.create(listener, presenter);
+    interactor =
+        TestOffGameInteractor.create(playerOne, playerTwo, listener, presenter, scoreStream);
   }
 }
