+++ android/tutorials/tutorial3/build.gradle
@@ -54,7 +54,6 @@ dependencies {
     implementation deps.external.dagger
     implementation deps.external.rxbinding
     implementation deps.androidx.percent
-    implementation deps.external.guavaAndroid
     compileOnly deps.apt.javax
     testImplementation project(":libraries:rib-test")
 }
