+++ android/tutorials/tutorial3/src/main/java/com/uber/rib/root/logged_in/off_game/OffGameInteractor.java
@@ -16,39 +16,25 @@
 package com.uber.rib.root.logged_in.off_game;
 
 import androidx.annotation.Nullable;
-import com.google.common.collect.ImmutableMap;
-import com.uber.autodispose.AutoDispose;
 import com.uber.rib.core.Bundle;
 import com.uber.rib.core.Interactor;
 import com.uber.rib.core.RibInteractor;
-import com.uber.rib.root.logged_in.ScoreStream;
 import io.reactivex.Observable;
 import io.reactivex.functions.Consumer;
 import javax.inject.Inject;
-import javax.inject.Named;
 
 /** Coordinates Business Logic for {@link OffGameScope}. */
 @RibInteractor
 public class OffGameInteractor
     extends Interactor<OffGameInteractor.OffGamePresenter, OffGameRouter> {
 
-  @Inject
-  @Named("player_one")
-  String playerOne;
-
-  @Inject
-  @Named("player_two")
-  String playerTwo;
-
   @Inject Listener listener;
   @Inject OffGamePresenter presenter;
-  @Inject ScoreStream scoreStream;
 
   @Override
   protected void didBecomeActive(@Nullable Bundle savedInstanceState) {
     super.didBecomeActive(savedInstanceState);
 
-    presenter.setPlayerNames(playerOne, playerTwo);
     presenter
         .startGameRequest()
         .subscribe(
@@ -56,19 +42,6 @@ public class OffGameInteractor
               @Override
               public void accept(Object object) throws Exception {
                 listener.onStartGame();
-              }
-            });
-
-    scoreStream
-        .scores()
-        .as(AutoDispose.<ImmutableMap<String, Integer>>autoDisposable(this))
-        .subscribe(
-            new Consumer<ImmutableMap<String, Integer>>() {
-              @Override
-              public void accept(ImmutableMap<String, Integer> scores) throws Exception {
-                Integer playerOneScore = scores.get(playerOne);
-                Integer playerTwoScore = scores.get(playerTwo);
-                presenter.setScores(playerOneScore, playerTwoScore);
               }
             });
   }
@@ -80,10 +53,6 @@ public class OffGameInteractor
 
   /** Presenter interface implemented by this RIB's view. */
   interface OffGamePresenter {
-
-    void setPlayerNames(String playerOne, String playerTwo);
-
-    void setScores(Integer playerOneScore, Integer playerTwoScore);
 
     Observable<Object> startGameRequest();
   }
