+++ android/tutorials/tutorial3/src/main/java/com/uber/rib/root/logged_in/tic_tac_toe/TicTacToeBuilder.java
@@ -26,7 +26,6 @@ import dagger.Binds;
 import dagger.BindsInstance;
 import dagger.Provides;
 import java.lang.annotation.Retention;
-import javax.inject.Named;
 import javax.inject.Qualifier;
 import javax.inject.Scope;
 
@@ -62,14 +61,7 @@ public class TicTacToeBuilder
   }
 
   public interface ParentComponent {
-
-    TicTacToeInteractor.Listener ticTacToeListener();
-
-    @Named("player_one")
-    String playerOne();
-
-    @Named("player_two")
-    String playerTwo();
+    // TODO: Define dependencies required from your parent interactor here.
   }
 
   @dagger.Module
