+++ android/tutorials/tutorial3/src/main/res/layout/logged_out_rib.xml
@@ -4,28 +4,22 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical">
-
     <Space
         android:layout_width="match_parent"
         android:layout_height="0dp"
         android:layout_weight="1"/>
-
     <EditText
-      android:id="@+id/player_one_name"
-      android:hint="@string/player_one_name"
-      android:layout_width="match_parent"
-      android:layout_height="wrap_content"
-      android:maxLines="1"
-      android:layout_margin="16dp"/>
-
+        android:id="@+id/player_name_1"
+        android:hint="Player Name 1"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_margin="16dp"/>
     <EditText
-      android:id="@+id/player_two_name"
-      android:hint="@string/player_two_name"
-      android:layout_width="match_parent"
-      android:layout_height="wrap_content"
-      android:maxLines="1"
-      android:layout_margin="16dp"/>
-
+        android:id="@+id/player_name_2"
+        android:hint="Player Name 2"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_margin="16dp"/>
     <Button
         android:id="@+id/login_button"
         android:text="Login"
