+++ android/tutorials/tutorial3/src/main/res/layout/off_game_rib.xml
@@ -7,48 +7,48 @@
     android:orientation="vertical">
 
     <LinearLayout
-      android:layout_width="match_parent"
-      android:layout_height="wrap_content"
-      android:layout_gravity="center_horizontal"
-      android:layout_margin="16dp"
-      android:orientation="horizontal">
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center_horizontal"
+        android:layout_margin="16dp"
+        android:orientation="horizontal">
 
         <TextView
-          android:id="@+id/player_one_name"
-          android:text="@string/player_one_name"
-          android:layout_width="0dp"
-          android:layout_height="wrap_content"
-          android:layout_weight="3"/>
+            android:id="@+id/player_one_name"
+            android:text="@string/player_one_name"
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            android:layout_weight="3"/>
 
         <TextView
-          android:id="@+id/player_one_win_count"
-          android:layout_width="0dp"
-          android:layout_height="wrap_content"
-          android:layout_weight="1"
-          android:text="@string/win_count"/>
+            android:id="@+id/player_one_win_count"
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            android:layout_weight="1"
+            android:text="@string/win_count"/>
 
     </LinearLayout>
 
     <LinearLayout
-      android:layout_width="match_parent"
-      android:layout_height="wrap_content"
-      android:orientation="horizontal"
-      android:layout_gravity="center_horizontal"
-      android:layout_margin="16dp">
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="horizontal"
+        android:layout_gravity="center_horizontal"
+        android:layout_margin="16dp">
 
         <TextView
-          android:id="@+id/player_two_name"
-          android:text="@string/player_two_name"
-          android:layout_width="0dp"
-          android:layout_height="wrap_content"
-          android:layout_weight="3"/>
+            android:id="@+id/player_two_name"
+            android:text="@string/player_two_name"
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            android:layout_weight="3"/>
 
         <TextView
-          android:id="@+id/player_two_win_count"
-          android:layout_width="0dp"
-          android:layout_height="wrap_content"
-          android:layout_weight="1"
-          android:text="@string/win_count"/>
+            android:id="@+id/player_two_win_count"
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            android:layout_weight="1"
+            android:text="@string/win_count"/>
 
     </LinearLayout>
     <Button
