+++ android/tutorials/tutorial3/src/test/java/com/uber/rib/root/logged_in/LoggedInInteractorTest.java
@@ -25,7 +25,6 @@ public class LoggedInInteractorTest extends RibTestBasePlaceholder {
 
   @Mock EmptyPresenter presenter;
   @Mock LoggedInRouter router;
-  @Mock MutableScoreStream scoreStream;
 
   private LoggedInInteractor interactor;
 
@@ -33,6 +32,6 @@ public class LoggedInInteractorTest extends RibTestBasePlaceholder {
   public void setup() {
     MockitoAnnotations.initMocks(this);
 
-    interactor = TestLoggedInInteractor.create(scoreStream);
+    interactor = TestLoggedInInteractor.create();
   }
 }
