+++ android/tutorials/tutorial3/src/test/java/com/uber/rib/root/logged_in/tic_tac_toe/TicTacToeInteractorTest.java
@@ -22,13 +22,9 @@ import org.mockito.MockitoAnnotations;
 
 public class TicTacToeInteractorTest extends RibTestBasePlaceholder {
 
-  private final String playerOne = "playerOne";
-  private final String playerTwo = "playerTwo";
-
-  @Mock Board board;
-  @Mock TicTacToeInteractor.Listener listener;
   @Mock TicTacToeInteractor.TicTacToePresenter presenter;
   @Mock TicTacToeRouter router;
+  @Mock Board board;
 
   private TicTacToeInteractor interactor;
 
@@ -36,6 +32,6 @@ public class TicTacToeInteractorTest extends RibTestBasePlaceholder {
   public void setup() {
     MockitoAnnotations.initMocks(this);
 
-    interactor = TestTicTacToeInteractor.create(board, listener, presenter, playerOne, playerTwo);
+    interactor = TestTicTacToeInteractor.create(board, presenter);
   }
 }
