+++ android/tutorials/tutorial3/src/test/java/com/uber/rib/root/logged_out/LoggedOutInteractorTest.java
@@ -46,21 +46,19 @@ public class LoggedOutInteractorTest extends RibTestBasePlaceholder {
 
   @Test
   public void attach_whenViewEmitsName_shouldCallListener() {
-    String fakeName1 = "1";
-    String fakeName2 = "2";
-    when(presenter.playerNames()).thenReturn(Observable.just(new Pair<>(fakeName1, fakeName2)));
+    when(presenter.loginName()).thenReturn(Observable.just(Pair.create("fakename", "fakename")));
 
     InteractorHelper.attach(interactor, presenter, router, null);
 
-    verify(listener).requestLogin(any(String.class), any(String.class));
+    verify(listener).login(any(String.class), any(String.class));
   }
 
   @Test
   public void attach_whenViewEmitsEmptyName_shouldNotCallListener() {
-    when(presenter.playerNames()).thenReturn(Observable.just(new Pair<String, String>("", "")));
+    when(presenter.loginName()).thenReturn(Observable.just(Pair.create("", "")));
 
     InteractorHelper.attach(interactor, presenter, router, null);
 
-    verify(listener, never()).requestLogin(any(String.class), any(String.class));
+    verify(listener, never()).login(any(String.class), any(String.class));
   }
 }
